Another approach to setup libvirt network

+++ config/install.sh
@@ -231,13 +231,11 @@ chown root:root /lib/udev/rules.d/60-qemu-system-common.rules
 
 # Setup libvirt network
 cp gns3.xml /etc/libvirt/qemu/networks/gns3.xml
-set +e
-virsh net-destroy default
-virsh net-undefine default
-virsh net-define /etc/libvirt/qemu/networks/gns3.xml
-virsh net-start gns3
-virsh net-autostart gns3
-set -e
+if virsh net-info default | grep -q '^Active:.*yes'; then
+    virsh net-destroy default
+    virsh net-define /etc/libvirt/qemu/networks/gns3.xml
+    virsh net-autostart gns3
+fi
 
 # Setup Console
 cp "console-setup" "/etc/default/console-setup"
